src/MCP/search/requirements.txt
httpx
uvicorn
mcp
python-dotenv==1.0.0
azure-search-documents
azure-identity==1.17.1
